+++ src/relstorage/adapters/mysql/procs/hp/lock_and_choose_tid_p.sql
@@ -10,6 +10,11 @@ BEGIN
     DECLARE scratch BIGINT;
     DECLARE current_tid_64 BIGINT;
 
+    -- We're in the commit phase of two-phase commit.
+    -- It's very important not to error out here.
+    -- So we need a very long wait to get the commit lock.
+    SET SESSION innodb_lock_wait_timeout = 500;
+
     SELECT tid
     INTO scratch
     FROM commit_row_lock